+++ languages/index.php
@@ -1,8 +1,9 @@
 <?php
 /**
- * Posts in Sidebar plugin index.php plugin.
+ * Posts in Sidebar plugin index.php file
  *
  * @package PostsInSidebar
+ * @since 2.0.0
  */
 
 /* Silence is golden */